Fix wrong size of event counter in BluetoothCccHandlerCallback
Bug: 448097575
Test: make

+++ bluetooth/bluetooth_hal/extensions/ccc/bluetooth_ccc.cc
@@ -122,7 +122,7 @@ class BluetoothCccHandlerCallbackImpl : public BluetoothCccHandlerCallback {
   void OnEventGenerated(const CccTimestamp& timestamp,
                         const BluetoothAddress& address, CccDirection direction,
                         CccLmpEventId lmp_event_id,
-                        uint8_t event_counter) override {
+                        uint16_t event_counter) override {
     if (bluetooth_ccc_callback_ == nullptr) {
       return;
     }
